added get functions for tag and category services

+++ src/main/java/com/yuzarsif/freelance/request/CreateAdvertisementRequest.java
@@ -3,7 +3,7 @@ package com.yuzarsif.freelance.request;
 public record CreateAdvertisementRequest(
         String header,
         String description,
-        int[] categoryId,
+        Long[] categoryId,
         Long[] tagId,
         Long employerId,
         int price,
